requirements.txt
tqdm==4.67.1
numpy>=1.26.0
pandas==2.2.3
jiwer==3.1.0
transformers>=4.43.0
torch==2.6.0
torchvision
speechbrain==1.0.2
librosa==0.10.2.post1

# utils
ipython
coqui-tts>=0.25.2
openai-whisper
ffmpeg